+++ src/main/java/io/skymind/pathmind/ui/views/experiment/RewardFunctionView.java
@@ -30,7 +30,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 @StyleSheet("frontend://styles/styles.css")
 @Route(value = "rewardFunction", layout = MainLayout.class)
-public class RewardFunctionView extends VerticalLayout implements BasicViewInterface, HasUrlParameter<Integer>
+public class RewardFunctionView extends VerticalLayout implements BasicViewInterface, HasUrlParameter<Long>
 {
 	private Logger log = LogManager.getLogger(RewardFunctionView.class);
 
@@ -124,7 +124,7 @@ public class RewardFunctionView extends VerticalLayout implements BasicViewInter
 
 	// TODO -> Duplicate code for now as I refactor the code on my new understanding from the call yesterday.
 	@Override
-	public void setParameter(BeforeEvent event, Integer projectId) {
+	public void setParameter(BeforeEvent event, Long projectId) {
 		Project project = projectRepository.getProject(projectId);
 		if(project == null) {
 			log.info("TODO -> Implement");
